Add support for interactivity
This commit makes the question/submit tutorial interaction function. I
suspect we'll run into some odd corner cases, but for now this is at
least a working implementation.

The tutorial now keeps track of section state and whether or not its
waiting for an answer to a question. If a given section does not need to
wait for an answer we proceed automatically to the next section.

Here's an example:

```
(tutorial)
=> # Introduction

   Welcome to my tutorial! Submit true to continue
(submit true)
=> Great job!

   You've reached the end of the tutorial, congrats!
```